feat: added frontend endpoints
these endpoints works when the build of the frontend through outpout : 'export' next config is copied to the
resources directory with static directory name. I do have included this directory to the git ignore
becouse I do not want to include a temporary build in the frontend project. I do not know yet what sould
be the best option to store this build.