fix: fixed and issue with the docker container
The issue was that the last stage of the Dockerimage was copying and pasteing the frontend build to the app directory.
But the jar had no access to these resources. So I have copy pasted the frontend build before createing the jar file.
In this way the jar file already has the necessery static resources. Now building this container takes more time since the
backend stage has to wait for the frontend page and cannot build seperatly. I think there is a way to copy and paste into the jar
file but I have no idea how to do it inisde the Dockerfile.