feat: adedd animal properties
I have added breed and size properties in orded to determain the length of the grooming.
I have implemented as an enum and connected these enums to the dog entity.
Also created an endpoint which responsible for adding new dogs.
The value of these enums are parsed when recived or sending from/to the frontend.